python/requirements.txt
django==4.1
django-cors-headers==3.10.0
djangorestframework==3.14
tzdata==2023.3
django-environ==0.11.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
django-filter==23.5
ipython==7.28.0